For CSS, make diagram the paren node of all items
Now I can change background color based on elements in the diagram:

    diagram:has(classitem) {
     background-color: beige;
    }
